android: binder: More offset validation.
Make sure offsets don't point to overlapping flat_binder_object
structs.

+++ drivers/android/binder.c
@@ -1516,10 +1516,10 @@ static void binder_transaction(struct binder_proc *proc,
 		    t->buffer->data_size < sizeof(*fp) ||
 		    !IS_ALIGNED(*offp, sizeof(u32))) {
 			binder_user_error("%d:%d got transaction with invalid offset, %lld (min %lld, max %lld)\n",
-			  proc->pid, thread->pid, (u64)*offp,
-			  off_min,
-			  (t->buffer->data_size -
-			     sizeof(*fp)));
+					  proc->pid, thread->pid, (u64)*offp,
+					  (u64)off_min,
+					  (u64)(t->buffer->data_size -
+					  sizeof(*fp)));
 			return_error = BR_FAILED_REPLY;
 			goto err_bad_offset;
 		}
